+++ src/hooks/useLocalStorage.ts
@@ -3,35 +3,44 @@ import { useCallback, useMemo } from 'react';
 /**
  * Local Storage Model
  */
-export const useLocalStorage = (key: string) => {
-  const get = useCallback(() => {
-    return window.localStorage.getItem(key);
-  }, [key]);
+export const useLocalStorage = (key: string, version: number) => {
+  const get = useCallback(async () => {
+    const json = window.localStorage.getItem(key);
+    if (json) {
+      const parsed = JSON.parse(json);
+
+      if (parsed?.version === version) {
+        return parsed.data;
+      }
+
+      return undefined;
+    }
+
+    return undefined;
+  }, [key, version]);
 
   /**
    * Update
    */
   const update = useCallback(
-    <T extends string>(data: T) => {
-      window.localStorage.setItem(key, data);
+    async <T>(data: T) => {
+      window.localStorage.setItem(
+        key,
+        JSON.stringify({
+          version,
+          data,
+        })
+      );
       return data;
     },
-    [key]
+    [key, version]
   );
-
-  /**
-   * Remove
-   */
-  const remove = useCallback(() => {
-    window.localStorage.removeItem(key);
-  }, [key]);
 
   return useMemo(
     () => ({
       get,
       update,
-      remove,
     }),
-    [get, update, remove]
+    [get, update]
   );
 };