+++ src/module.ts
@@ -3,6 +3,7 @@ import { getTemplateSrv } from '@grafana/runtime';
 
 import { GroupsEditor, VariablePanel } from './components';
 import {
+  ALLOW_CUSTOM_VALUE_OPTIONS,
   ALLOW_EMPTY_VALUE_OPTIONS,
   ALWAYS_VISIBLE_FILTER_OPTIONS,
   AUTO_SCROLL_OPTIONS,
@@ -112,6 +113,16 @@ export const plugin = new PanelPlugin<PanelOptions>(VariablePanel)
           options: SHOW_LABEL_OPTIONS,
         },
         showIf: (config) => showForButtonView(config),
+      })
+      .addRadio({
+        path: 'customValue',
+        name: 'Allow custom values',
+        description: 'Supports custom values for the variable.',
+        defaultValue: false,
+        settings: {
+          options: ALLOW_CUSTOM_VALUE_OPTIONS,
+        },
+        showIf: (config) => showForMinimizeView(config),
       });
 
     builder
@@ -221,15 +232,17 @@ export const plugin = new PanelPlugin<PanelOptions>(VariablePanel)
         defaultValue: false,
         category: ['Layout'],
         showIf: (config) => showForTableView(config) && !!config.groups?.length,
-      })
+      });
+
+    builder
       .addRadio({
         path: 'groupSelection',
-        name: 'Allow group selection.',
+        name: 'Allow group selection',
         settings: {
           options: GROUP_SELECTION_OPTIONS,
         },
         defaultValue: false,
-        category: ['Layout'],
+        category: ['Groups'],
         showIf: (config) => showForTableView(config) && !!config.groups?.length,
       })
       .addRadio({
@@ -240,8 +253,19 @@ export const plugin = new PanelPlugin<PanelOptions>(VariablePanel)
           options: SELECTED_GROUP_OPTIONS,
         },
         defaultValue: false,
-        category: ['Layout'],
+        category: ['Groups'],
         showIf: (config) => showForTableView(config) && config.groups?.length > 1,
+      })
+      .addTextInput({
+        path: 'saveSelectedGroupKey',
+        name: 'Selected group Id',
+        description: 'Key to preserve selected group in the browser storage.',
+        defaultValue: '',
+        settings: {
+          placeholder: 'Will be unique per panel if empty',
+        },
+        showIf: (config) => showForTableView(config) && config.saveSelectedGroup,
+        category: ['Groups'],
       });
 
     builder.addSelect({