+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Added the indeterminate state to the checkboxes (#127)
 - Updated Tree View scrolling (#128)
 - Add Responsive tabs (#130)
+- Add support new items in Variable (minimize mode) (#134)
+- Add Preserve selection of the group between dashboards (#135)
 
 ## 2.4.0 (2024-03-17)